correcciones codigo
