+++ Src/NetApiRaidenTemplate.Wizard/Resources/TemplateFiles/Readme.html
@@ -94,8 +94,16 @@
         </header>
 
         <section>
+			/* $identityserver_feature$ start */
             <article>
-                <h3></h3>
+                <h3>IdentityServer:</h3>
+                <p>Ejecutar los siguientes comandos para generar las migraciones:</p>
+				<p>Ejecutar los siguientes comandos para generar las migraciones:</p> 
+            </article>
+			/* $identityserver_feature$ end */
+			
+			<article>
+                <h3>TODO:</h3>
                 <p></p>
                 <address></address>
             </article>
@@ -105,7 +113,7 @@
     </div>
 
     <footer>
-        <span>© Copyright David González Zornoza 2021</span>
+        <span>© Copyright David González Zornoza 2022</span>
         <span class="pull-right"><a href="https://github.com/dgzornoza">Github</a></span>
     </footer>
 </body>
