+++ WebApplication1/Src/WebApplication1.API/Controllers/AuthenticateController.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using IdentityServer4;
