+++ WebApplication1/Src/WebApplication1.API/Startup.cs
@@ -1,21 +1,9 @@
-﻿/* $identityserver_feature$ start */
-using System.IdentityModel.Tokens.Jwt;
-using System.Reflection;
-using System.Text;
-using IdentityServer4.Configuration;
-using Microsoft.AspNetCore.Authentication.JwtBearer;
-/* $identityserver_feature$ end */
-using Microsoft.AspNetCore.Builder;
+﻿using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Mvc.ApiExplorer;
-using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
-/* $identityserver_feature$ start */
-using Microsoft.IdentityModel.Tokens;
-using WebApplication1.Api.Infrastructure.Authorization;
-/* $identityserver_feature$ end */
 using WebApplication1.Api.Infrastructure.Extensions;
 using WebApplication1.Api.Infrastructure.Filters;
 
@@ -55,47 +43,6 @@ namespace WebApplication1.Api
             services.AddAppConfiguration(this.Configuration);
 
             services.AddIocContainer(this.Configuration);
-
-            /* $identityserver_feature$ start */
-            ////services.AddIdentityServer(options =>
-            ////{
-            ////    ////options.UserInteraction = new UserInteractionOptions()
-            ////    ////{
-            ////    ////    LogoutUrl = "set logout url",
-            ////    ////    LoginUrl = "set login url",
-            ////    ////    LoginReturnUrlParameter = "returnUrl",
-            ////    ////};
-            ////})
-            ////.AddInMemoryClients(IdentityConfiguration.Clients)
-            ////.AddInMemoryIdentityResources(IdentityConfiguration.IdentityResources)
-            ////.AddInMemoryApiResources(IdentityConfiguration.ApiResources)
-            ////.AddInMemoryApiScopes(IdentityConfiguration.ApiScopes)
-            ////.AddTestUsers(IdentityConfiguration.TestUsers)
-            ////.AddDeveloperSigningCredential();
-
-            ////services.AddAuthentication(options =>
-            ////{
-            ////    options.DefaultAuthenticateScheme = JwtBearerDefaults.AuthenticationScheme;
-            ////    options.DefaultChallengeScheme = JwtBearerDefaults.AuthenticationScheme;
-            ////    options.DefaultScheme = JwtBearerDefaults.AuthenticationScheme;
-            ////})
-            ////.AddJwtBearer(options =>
-            ////{
-            ////    options.SaveToken = true;
-            ////    options.RequireHttpsMetadata = false;
-            ////    options.Authority = this.Configuration["JWT:Authority"];
-            ////    options.Audience = this.Configuration["JWT:ValidAudience"];
-
-            ////    options.TokenValidationParameters = new TokenValidationParameters()
-            ////    {
-            ////        ValidateIssuer = true,
-            ////        ValidateAudience = true,
-            ////        ValidAudience = this.Configuration["JWT:ValidAudience"],
-            ////        ValidIssuer = this.Configuration["JWT:ValidIssuer"],
-            ////        IssuerSigningKey = new SymmetricSecurityKey(Encoding.UTF8.GetBytes(this.Configuration["JWT:Secret"])),
-            ////    };
-            ////});
-            /* $identityserver_feature$ end */
         }
 
         /// <summary>
